python: Add blinkonprivate add-on
This simple Python add-on, when activated, will blink the tray icon
whenever the user's nick is mentioned in a private window (aka query).

Useful in case ithe user wants the alert only on nick mentions but not
in every message received in private window.

Special thanks to @rrosatti for the help with testing.
Also, thanks to authors of nicenicks and statusmsg add-ons - I've studied
their code to learn how to do this one.

Signed-off-by: Guilherme G. Piccoli <gpiccoli@linux.vnet.ibm.com>